server/requirements.txt
asgiref==3.8.1
cachetools==5.5.0
certifi==2024.8.30
cffi==1.17.1
chardet==5.2.0
charset-normalizer==3.3.2
click==8.1.7
cloudinary==1.41.0
colorama==0.4.6
crispy-bootstrap5==2024.2
cryptography==43.0.1
distlib==0.3.8
Django==5.1.1
django-admin-honeypot==1.1.0
django-cloudinary-storage==0.3.0
django-cors-headers==4.4.0
django-crispy-forms==2.3
django-daraja==1.3.0
django-environ==0.11.2
django-jazzmin==3.0.0
django-livereload-server==0.5.1
django-otp==1.5.4
django-rest-framework==0.1.0
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
filelock==3.16.0
gunicorn==23.0.0
h11==0.14.0
idna==3.8
packaging==24.1
pip-review==1.3.0
platformdirs==4.3.2
pluggy==1.5.0
psycopg2-binary==2.9.9
pycparser==2.22
PyJWT==2.9.0
pymemcache==4.0.0
pyproject-api==1.7.1
python-decouple==3.8
python-dotenv==1.0.1
requests==2.32.3
rest-framework-simplejwt==0.0.2
setuptools==74.1.2
six==1.16.0
sqlparse==0.5.1
tomli==2.0.1
tornado==6.4.1
tox==4.18.1
typing_extensions==4.12.2
tzdata==2024.1
urllib3==2.2.2
uvicorn==0.30.6
virtualenv==20.26.4
wheel==0.44.0
whitenoise==6.7.0
